socket implemented and the frontend and backend are connected .

+++ backend/src/utils/cloudinary.js
@@ -17,10 +17,15 @@ const uploadonCloudinary= async (localFilePath) => {
         })
 
         
-            fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
             return response;
     }catch(error){
-        fs.unlinkSync(localFilePath)
+       
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         return null;
     }
 }
